PrakharAndSagar_GenAiOps/xpia/requirements.txt
crewai
fastapi
uvicorn
supabase
python-dotenv
requests
beautifulsoup4
selenium
webdriver-manager
pydantic
google-generativeai
langchain-google-genai
gunicorn